requirements.txt
mesop>=0.4.0
pandas>=2.0.0
numpy>=1.24.0
#sap-rpt-oss
huggingface_hub>=0.20.0
plotly>=5.17.0
requests>=2.31.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
gradio
pyarrow>=10.0.0
pyzmq>=25.0.0

